Split planner DSL return body and case helpers
- Move case DSL helpers into subject-family child modules behind the existing facade.
- Move function return-body DSL helpers into primitive, function, and conversion modules.
- Keep planner::dsl import surface stable through parent facade re-exports.
- Colocate DSL unit tests with the helper modules that own each shape.
